Copied cinema project to split off 2K and 4K versions. refs TweetWallFX/TweetwallFX-Devoxx-2018-BE#13
