+++ cinema2k/build.gradle
@@ -1,0 +1,47 @@
+/*
+ * The MIT License
+ *
+ * Copyright 2018 TweetWallFX
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+apply plugin: 'java-library'
+apply plugin: 'application'
+mainClassName = 'org.tweetwallfx.generic.Main'
+
+tasks.withType(JavaExec) {
+    main mainClassName
+    jvmArgs '-Dorg.tweetwallfx.scheduledata.day=tuesday'
+    jvmArgs '-Dorg.tweetwallfx.scheduledata.time=10:00Z'
+    jvmArgs '-Dorg.tweetwallfx.scheduledata.proposal=CTH-5247'
+}
+
+task('debug', dependsOn: 'classes', type: JavaExec) {
+    classpath sourceSets.main.runtimeClasspath
+    debug = true
+}
+
+dependencies {
+    implementation project(':tweetwallfx-devoxx-2018-be-stepengine')
+    implementation project(':tweetwallfx-devoxx-2018-be-cfp-impl')
+    implementation group: 'org.tweetwallfx', name: 'tweetwallfx-tweet-impl-twitter4j', version: versionTweetwallFX
+    implementation group: 'org.tweetwallfx', name: 'tweetwallfx-generic2d', version: versionTweetwallFX
+    implementation group: 'org.apache.logging.log4j', name: 'log4j-core', version: versionLog4j
+}
