+++ settings.gradle
@@ -32,5 +32,7 @@ project(':stepengine').name = 'tweetwallfx-devoxx-2018-be-stepengine'
 
 include ':exhibition'
 project(':exhibition').name = 'tweetwallfx-devoxx-2018-be-exhibition'
-include ':cinema'
-project(':cinema').name = 'tweetwallfx-devoxx-2018-be-cinema'
+include ':cinema2k'
+project(':cinema2k').name = 'tweetwallfx-devoxx-2018-be-cinema2k'
+include ':cinema4k'
+project(':cinema4k').name = 'tweetwallfx-devoxx-2018-be-cinema4k'
